+++ src/optical_dispersion_relations/plasmon.py
@@ -37,7 +37,7 @@ def metal_insulator_metal_collin_approximation(dielectric_permittivity: float,
         insulator_thickness, in the same unit of distance as wavelength: float
 
     Returns:
-        effective_refractive_index of the light propagating in the wavevuide: complex
+        effective_refractive_index of the light propagating in the waveguide: complex
 
     Derivation in:
         Waveguiding in nanoscale metallic apertures.
@@ -52,9 +52,9 @@ def metal_insulator_metal_collin_approximation(dielectric_permittivity: float,
 
 
 def metal_insulator_metal_sondergaard_narrow_approximation(dielectric_permittivity: float,
-                                                    metal_permittivity: complex,
-                                                    wavelength: float,
-                                                    insulator_thickness: float) -> complex:
+                                                           metal_permittivity: complex,
+                                                           wavelength: float,
+                                                           insulator_thickness: float) -> complex:
     '''Approximate metal-insulator-metal waveguide dispersion relation for TM polarization.
 
     Parameters: