+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+* Documentation for:
+    * `metal_insulator_metal_collin_approximation`
+    * `metal_insulator_metal_sondergaard_narrow_approximation`
+
+## Added
+
+* Example figures and scripts under `./examples`
+
 ## [v0.0.0](https://github.com/g-duff/optical_dispersion_relations/releases/v0.0.0)
 
 ### Added
@@ -25,5 +35,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * [src-layout](https://setuptools.pypa.io/en/latest/userguide/package_discovery.html#src-layout) package directory structure.
-
-### Removed